remove old FilesPageFragment
