+++ app/src/main/java/net/yupol/transmissionremote/app/torrentdetails/FilesPageFragment.java
@@ -2,47 +2,71 @@ package net.yupol.transmissionremote.app.torrentdetails;
 
 import android.content.Context;
 import android.os.Bundle;
+import android.support.annotation.AnimRes;
+import android.support.annotation.Nullable;
+import android.support.v4.app.FragmentManager;
+import android.support.v4.app.FragmentTransaction;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.CompoundButton;
-import android.widget.ListView;
-import android.widget.ProgressBar;
 
+import net.yupol.transmissionremote.app.ProgressbarFragment;
 import net.yupol.transmissionremote.app.R;
-import net.yupol.transmissionremote.app.model.json.FileStat;
-import net.yupol.transmissionremote.app.model.json.Torrent;
+import net.yupol.transmissionremote.app.model.Dir;
 import net.yupol.transmissionremote.app.model.json.TorrentInfo;
 import net.yupol.transmissionremote.app.transport.BaseSpiceActivity;
-import net.yupol.transmissionremote.app.transport.TransportManager;
-import net.yupol.transmissionremote.app.transport.request.TorrentSetRequest;
 
-public class FilesPageFragment extends BasePageFragment {
+import java.util.ArrayList;
+import java.util.Stack;
+
+public class FilesPageFragment extends BasePageFragment implements DirectoryFragment.OnDirectorySelectedListener {
 
     private static final String TAG = FilesPageFragment.class.getSimpleName();
-
-    private ListView list;
-    private ProgressBar progressBar;
-    private TransportManager transportManager;
+    private static final String TAG_PROGRESSBAR_FRAGMENT = "tag_progressbar_fragment";
+    private static final String TAG_DIRECTORY_FRAGMENT = "tag_directory_fragment";
+    private static final String KEY_PATH = "key_path";
 
     private boolean viewCreated;
+    private Stack<Dir> path = new Stack<>();
+    private BreadcrumbView breadcrumbView;
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         super.onCreateView(inflater, container, savedInstanceState);
 
-        View view = inflater.inflate(R.layout.torrent_details_file_page_fragment, container, false);
-        list = (ListView) view.findViewById(R.id.file_list);
-        progressBar = (ProgressBar) view.findViewById(R.id.progress_bar);
-        list.setVisibility(View.GONE);
-        progressBar.setVisibility(View.VISIBLE);
+        View view = inflater.inflate(R.layout.torrent_details_file_page_fragment2, container, false);
+        breadcrumbView = (BreadcrumbView) view.findViewById(R.id.breadcrumb_view);
+        breadcrumbView.setOnNodeSelectedListener(new BreadcrumbView.OnNodeSelectedListener() {
+            @Override
+            public void onNodeSelected(int position) {
+                if (position >= path.size() - 1) return;
+                for (int i=path.size()-1; i>position; i--) {
+                    path.remove(i);
+                }
+                breadcrumbView.setPath(path);
+                showDirectory(path.peek(), AnimationDirection.LEFT_TO_RIGHT);
+            }
+        });
+
+        if (savedInstanceState != null) {
+            ArrayList<Dir> dirs = savedInstanceState.getParcelableArrayList(KEY_PATH);
+            path.addAll(dirs);
+            breadcrumbView.setPath(path);
+        }
 
         if (getTorrentInfo() != null) {
-            showList();
+            if (path.isEmpty()) {
+                showRootDir();
+            } else {
+                showDirectory(path.peek(), null);
+            }
+        } else {
+            showProgressbarFragment();
         }
 
         viewCreated = true;
+
 
         return view;
     }
@@ -58,52 +82,83 @@ public class FilesPageFragment extends BasePageFragment {
         super.onAttach(context);
         if (!(getActivity() instanceof BaseSpiceActivity)) {
             Log.e(TAG, "Fragment should be used with BaseSpiceActivity");
-            return;
         }
-        transportManager = ((BaseSpiceActivity) getActivity()).getTransportManager();
     }
+
+    @Override
+    public void onSaveInstanceState(Bundle outState) {
+        super.onSaveInstanceState(outState);
+        outState.putParcelableArrayList(KEY_PATH, new ArrayList<>(path));
+    }
+
 
     @Override
     public void setTorrentInfo(TorrentInfo torrentInfo) {
         super.setTorrentInfo(torrentInfo);
         if (viewCreated) {
-            showList();
+            showRootDir();
         }
     }
 
-    private void showList() {
-        list.setAdapter(new FilesAdapter(getTorrentInfo(), new FileSelectedListener(getTorrent(), transportManager)));
-        list.setVisibility(View.VISIBLE);
-        progressBar.setVisibility(View.GONE);
+    @Override
+    public void onDirectorySelected(Dir dir) {
+        path.push(dir);
+        breadcrumbView.setPath(path);
+        showDirectory(dir, AnimationDirection.RIGHT_TO_LEFT);
     }
 
-    private static class FileSelectedListener implements CompoundButton.OnCheckedChangeListener {
-
-        private Torrent torrent;
-        private TransportManager transportManager;
-
-        public FileSelectedListener(Torrent torrent, TransportManager transportManager) {
-            this.torrent = torrent;
-            this.transportManager = transportManager;
+    @Override
+    public boolean onBackPressed() {
+        if (path.size() > 1) {
+            path.pop();
+            breadcrumbView.setPath(path);
+            showDirectory(path.peek(), AnimationDirection.LEFT_TO_RIGHT);
+            return true;
         }
+        return super.onBackPressed();
+    }
 
-        @Override
-        public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-            if (!(buttonView.getTag() instanceof FileStat)) return;
+    private void showRootDir() {
+        Dir rootDir = Dir.createFileTree(getTorrentInfo().getFiles());
+        showDirectory(rootDir, null);
+        path.clear();
+        path.push(rootDir);
+        breadcrumbView.setPath(path);
+    }
 
-            FileStat fileStat = (FileStat) buttonView.getTag();
-            if (fileStat.isWanted() != isChecked) {
-                fileStat.setWanted(isChecked);
+    private void showProgressbarFragment() {
+        FragmentManager fm = getChildFragmentManager();
+        ProgressbarFragment progressbarFragment = (ProgressbarFragment) fm.findFragmentByTag(TAG_PROGRESSBAR_FRAGMENT);
+        if (progressbarFragment == null) {
+            progressbarFragment = new ProgressbarFragment();
+            FragmentTransaction ft = fm.beginTransaction();
+            ft.replace(R.id.content, progressbarFragment, TAG_PROGRESSBAR_FRAGMENT);
+            ft.commitAllowingStateLoss();
+        }
+    }
 
-                int fileIndex = (int) buttonView.getTag(R.id.TAG_FILE_INDEX);
-                TorrentSetRequest.Builder requestBuilder = TorrentSetRequest.builder(torrent.getId());
-                if (fileStat.isWanted()) {
-                    requestBuilder.filesWanted(fileIndex);
-                } else {
-                    requestBuilder.filesUnwanted(fileIndex);
-                }
-                transportManager.doRequest(requestBuilder.build(), null);
-            }
+    private void showDirectory(Dir dir, @Nullable AnimationDirection animation) {
+        TorrentInfo torrentInfo = getTorrentInfo();
+        DirectoryFragment fragment = DirectoryFragment.newInstance(
+                getTorrent().getId(), dir, torrentInfo.getFiles(), torrentInfo.getFileStats());
+        FragmentTransaction ft = getChildFragmentManager().beginTransaction();
+        if (animation != null) {
+            ft.setCustomAnimations(animation.enter, animation.exit);
+        }
+        ft.replace(R.id.content, fragment, TAG_DIRECTORY_FRAGMENT);
+        ft.commitAllowingStateLoss();
+    }
+
+    private enum AnimationDirection {
+        RIGHT_TO_LEFT(R.anim.enter_from_right, R.anim.exit_to_left),
+        LEFT_TO_RIGHT(R.anim.enter_from_left, R.anim.exit_to_right);
+
+        @AnimRes public final int enter;
+        @AnimRes public final int exit;
+
+        AnimationDirection(@AnimRes int enter, @AnimRes int exit) {
+            this.enter = enter;
+            this.exit = exit;
         }
     }
 }
