+++ app/src/main/java/net/yupol/transmissionremote/app/torrentdetails/TorrentDetailsPagerAdapter.java
@@ -15,7 +15,6 @@ public class TorrentDetailsPagerAdapter extends FragmentStatePagerAdapter {
     private Class<?>[] fragmentsClasses = {
             TorrentInfoPageFragment.class,
             FilesPageFragment.class,
-            FilesPageFragment2.class,
             OptionsPageFragment.class,
     };
